regular service -  payment

+++ lib/Models/customer_models/mechanic_details_model/mechanicDetailsMdl.dart
@@ -187,7 +187,7 @@ class Mechanic {
   dynamic certificate1;
   dynamic certificate2;
   int status;
-  int rate;
+  double rate;
   int reviewCount;
 
   factory Mechanic.fromJson(Map<String, dynamic> json) => Mechanic(
@@ -245,7 +245,7 @@ class MechanicReviewsDatum {
 
   String id;
   int transType;
-  int rating;
+  double rating;
   String feedback;
   int bookingId;
   dynamic orderId;
